Move Repository interfaces and DateRange into Domain project
