+++ src/RuneSharper.Data/Repositories/CachedSnapshotRepostiory.cs
@@ -2,8 +2,9 @@
 using Microsoft.Extensions.Options;
 using RuneSharper.Data.Extensions;
 using RuneSharper.Domain.Entities.Snapshots;
+using RuneSharper.Domain.Helpers;
+using RuneSharper.Domain.Interfaces;
 using RuneSharper.Shared.Extensions;
-using RuneSharper.Shared.Helpers;
 using RuneSharper.Shared.Settings;
 
 namespace RuneSharper.Data.Repositories;
