+++ src/RuneSharper.Data/Specifications/SnapshotBetweenDateRangeSpecification.cs
@@ -1,6 +1,6 @@
 ﻿using RuneSharper.Domain.Entities.Snapshots;
+using RuneSharper.Domain.Helpers;
 using RuneSharper.Shared.Enums;
-using RuneSharper.Shared.Helpers;
 
 namespace RuneSharper.Data.Specifications;
