+++ src/RuneSharper.Services/Snapshots/ChangeAggregation/Strategies/BaseChangeStrategy.cs
@@ -1,6 +1,6 @@
-﻿using RuneSharper.Services.DateTimeProvider;
+﻿using RuneSharper.Domain.Helpers;
+using RuneSharper.Services.DateTimeProvider;
 using RuneSharper.Shared.Enums;
-using RuneSharper.Shared.Helpers;
 
 namespace RuneSharper.Services.Snapshots.ChangeAggregation.Strategies;
